Refactor block collection tests.
